requirements.txt
# ========================================================================
# COMPLETE PROJECT STRUCTURE FOR MNIST DEEP LEARNING CLASSIFICATION
# ========================================================================

opencv-python-headless>=4.8.0
matplotlib>=3.6.0
numpy>=1.24.0
pillow>=9.0.0
tensorflow>=2.13.0
bokeh>=2.4.0
torch>=1.13.0
torchvision>=0.14.0
tqdm>=4.64.0
ipywidgets>=8.0.0
livelossplot>=0.5.4
pytest>=7.2.0
pandas>=1.5.0
seaborn>=0.11.0
jupyter>=1.0.0
ipykernel>=6.20.0
scikit-learn>=1.3.0